+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 ### Fixes
 
 - Silence spurious ERROR log in `SentryCrashCxaThrowSwapper` for empty sections (#8915)
+- Stop recording touch events while Session Replay is paused. (#8887)
 
 ### Internal